herbie_vision/model/rest_api/requirements.txt
opencv-python
numpy
requests
flask
pillow
torch
torchvision
google-cloud-storage